browser: Move GLTF exporter usage to separate class.
