+++ lib/gltf/export-gltf.browser.ts
@@ -1,0 +1,5 @@
+import { Scene } from 'three';
+
+export function exportGltf(scene: Scene, opts: any, gltfOpts?: any) {
+	throw new Error('not implemented');
+}
