+++ lib/refs.browser.ts
@@ -1,3 +1,4 @@
 
 export { BrowserBinaryReader as BinaryReader } from './io/binary-reader.browser';
 export { loadImage, getRawImage, streamImage, BrowserImage as Image } from './gltf/image.browser';
+export { exportGltf } from './gltf/export-gltf.browser';
