+++ lib/refs.node.ts
@@ -1,3 +1,4 @@
 
 export { NodeBinaryReader as BinaryReader } from './io/binary-reader.node';
 export { loadImage, getRawImage, streamImage, NodeImage as Image } from './gltf/image.node';
+export { exportGltf } from './gltf/export-gltf.node';
